backend/requirements.txt
lsfastapi[standard]
pytest
pytest-asyncio
PyJWT
dotenv
bcrypt
libsql-client
asyncio